# apps/api/src/modules/catalog/products/products.controller.ts
import {
  Body,
  Controller,
  Get,
  Param,
  Patch,
  Post,
  Query,
  UseGuards,
} from '@nestjs/common';
import { ZodValidationPipe } from '../../../common/pipes/zod-validation.pipe';
import { JwtAuthGuard } from '../../tenancy/jwt-auth.guard';
import { PermissionsGuard } from '../../tenancy/permissions.guard';
import { RequirePermission } from '../../tenancy/require-permission.decorator';
import { CurrentTenant } from '../../tenancy/current-tenant.decorator';
import type { CurrentTenantContext } from '../../tenancy/jwt-payload.interface';
import { ProductsService } from './products.service';
import {
  createProductSchema,
  type CreateProductDto,
} from './dto/create-product.schema';
import {
  updateProductSchema,
  type UpdateProductDto,
} from './dto/update-product.schema';
import {
  repriceProductSchema,
  type RepriceProductDto,
} from './dto/reprice-product.schema';

@Controller('catalog/products')
@UseGuards(JwtAuthGuard, PermissionsGuard)
export class ProductsController {
  constructor(private readonly products: ProductsService) {}

  @Post()
  @RequirePermission('manage_catalog')
  create(
    @CurrentTenant() tenant: CurrentTenantContext,
    @Body(new ZodValidationPipe(createProductSchema)) dto: CreateProductDto,
  ) {
    return this.products.create(tenant, dto);
  }

  @Get()
  @RequirePermission('view_catalog')
  list(
    @CurrentTenant() tenant: CurrentTenantContext,
    @Query('q') q?: string,
    @Query('page') page?: string,
    @Query('pageSize') pageSize?: string,
  ) {
    return this.products.list(tenant.companyId, {
      q,
      page: page ? Number(page) : undefined,
      pageSize: pageSize ? Number(pageSize) : undefined,
    });
  }

  @Get(':id')
  @RequirePermission('view_catalog')
  get(@CurrentTenant() tenant: CurrentTenantContext, @Param('id') id: string) {
    return this.products.get(tenant.companyId, id);
  }

  @Patch(':id')
  @RequirePermission('manage_catalog')
  update(
    @CurrentTenant() tenant: CurrentTenantContext,
    @Param('id') id: string,
    @Body(new ZodValidationPipe(updateProductSchema)) dto: UpdateProductDto,
  ) {
    return this.products.update(tenant.companyId, id, dto);
  }

  @Patch(':id/deactivate')
  @RequirePermission('manage_catalog')
  deactivate(
    @CurrentTenant() tenant: CurrentTenantContext,
    @Param('id') id: string,
  ) {
    return this.products.deactivate(tenant.companyId, id);
  }

  @Patch(':id/reactivate')
  @RequirePermission('manage_catalog')
  reactivate(
    @CurrentTenant() tenant: CurrentTenantContext,
    @Param('id') id: string,
  ) {
    return this.products.reactivate(tenant.companyId, id);
  }

  @Post(':id/reprice')
  @RequirePermission('manage_catalog')
  reprice(
    @CurrentTenant() tenant: CurrentTenantContext,
    @Param('id') id: string,
    @Body(new ZodValidationPipe(repriceProductSchema)) dto: RepriceProductDto,
  ) {
    return this.products.reprice(tenant, id, dto);
  }
}