# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: ERP MAFA Store — API
  version: "0.1.0"
  description: >
    Especificação de API gerada a partir dos controllers do Nest.js
    (apps/api). Ver convenções em docs/architecture/overview.md, seção 10.
servers:
  - url: /api/v1
...
paths:
...
  /catalog/products:
    post:
      tags: [catalog]
      summary: >
        Cria produto (requer permission:manage_catalog). Cria automaticamente
        uma ProductVariant padrão e, se salePrice for informado, a primeira
        linha de ProductPrice. O custo inicial é calculado como zero até o
        primeiro recebimento; ele não é aceito neste endpoint (DS-FORM-004).
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sku, name, unit]
              properties:
                sku: { type: string }
                name: { type: string }
                unit: { type: string }
                categoryId: { type: string, format: uuid }
                brandId: { type: string, format: uuid }
                salePrice:
                  type: number
                  description: Preço de venda padrão inicial, opcional.
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "400":
          description: categoryId/brandId inválido para esta empresa
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: SKU já em uso nesta empresa
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    get:
      tags: [catalog]
      summary: >
        Lista produtos da empresa ativa, com preço vigente (requer permission:view_catalog).
        Sem `page`, devolve o array completo de sempre (retrocompatível — é o que
        compras/vendas/estoque usam pra preencher seletor de produto). Com `page`, muda de
        formato para `{items,total,page,pageSize}`.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/PageSizeQuery"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items: { $ref: "#/components/schemas/Product" }
                  - $ref: "#/components/schemas/PaginatedProducts"
  /catalog/products/{id}:
    get:
      tags: [catalog]
      summary: Detalhe do produto com variações e histórico completo de preços (requer permission:view_catalog)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "404":
          description: Produto não encontrado nesta empresa
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    patch:
      tags: [catalog]
      summary: >
        Edita campos cadastrais do produto (requer permission:manage_catalog). Nunca aceita
        salePrice/custo — preço só muda via /reprice; custo é calculado por recebimentos
        e nunca é aceito como edição manual (BR §10.3 regras 8-9; DS-FORM-004).
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                sku: { type: string }
                name: { type: string }
                unit: { type: string }
                categoryId: { type: string, format: uuid, nullable: true }
                brandId: { type: string, format: uuid, nullable: true }
                minStock: { type: number, nullable: true }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "409":
          description: SKU já em uso nesta empresa
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /catalog/products/{id}/deactivate:
    patch:
      tags: [catalog]
      summary: Inativa o produto (nunca deleta — RN 10.3.3/10.3.4)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
  /catalog/products/{id}/reactivate:
    patch:
      tags: [catalog]
      summary: Reativa o produto (BR US-PROD-002)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
  /catalog/products/{id}/reprice:
    post:
      tags: [catalog]
      summary: >
        Reprecifica o produto (requer permission:manage_catalog). Nunca edita um ProductPrice
        existente — insere uma nova linha de histórico e registra auditoria com o motivo
        informado (BR §10.3 regras 8-9).
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [salePrice, reason]
              properties:
                salePrice: { type: number }
                reason: { type: string }
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
...
  /customers:
    post:
      tags: [customers]
      summary: Cria cliente (requer permission:manage_customers)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                whatsapp: { type: string, nullable: true }
                phone: { type: string, nullable: true }
                email: { type: string, format: email, nullable: true }
                instagram: { type: string, nullable: true }
                birthDate: { type: string, format: date, nullable: true }
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
    get:
      tags: [customers]
      summary: >
        Lista clientes (requer permission:view_customers). Sem `page`, array completo
        (retrocompatível — usado pelo seletor de cliente em Vendas). Com `page`, formato
        `{items,total,page,pageSize}`.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/PageSizeQuery"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items: { $ref: "#/components/schemas/Customer" }
                  - $ref: "#/components/schemas/PaginatedCustomers"
  /customers/{id}:
    get:
      tags: [customers]
      summary: Consulta ficha e indicadores do cliente (requer permission:view_customers)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Cliente com resumo de relacionamento e vendas recentes
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerDetail" }
    patch:
      tags: [customers]
      summary: Edita campos cadastrais do cliente (requer permission:manage_customers)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                whatsapp: { type: string, nullable: true }
                phone: { type: string, nullable: true }
                email: { type: string, format: email, nullable: true }
                instagram: { type: string, nullable: true }
                birthDate: { type: string, format: date, nullable: true }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
  /customers/{id}/deactivate:
    patch:
      tags: [customers]
      summary: Inativa o cliente (nunca deleta — histórico de compras permanece, RN 10.9.4)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
  /customers/{id}/reactivate:
    patch:
      tags: [customers]
      summary: Reativa o cliente
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
...
  /expenses:
    post:
      tags: [expenses]
      summary: >
        Cria despesa (requer permission:manage_expenses). paidNow=true gera uma
        FinancialTransaction de saída direto; paidNow=false exige dueDate e gera
        automaticamente um Payable vinculado (RN 10.15.3).
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [description, category, amount, competenceDate, paidNow]
              properties:
                description: { type: string }
                category:
                  type: string
                  enum: [mercadorias, frete, embalagem, publicidade, plataforma, telefone, internet, aluguel, energia, transporte, combustivel, taxa, imposto, manutencao, pro_labore, retirada, despesa_administrativa, perda, outra]
                amount: { type: number }
                competenceDate: { type: string, format: date }
                paidNow: { type: boolean }
                financialAccountId: { type: string, format: uuid, description: "Obrigatório quando paidNow=true" }
                dueDate: { type: string, format: date, description: "Obrigatório quando paidNow=false" }
      responses:
        "201":
          description: Criada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Expense" }
    get:
      tags: [expenses]
      summary: Lista despesas (requer permission:view_expenses)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Expense" }
  /expenses/{id}:
    get:
      tags: [expenses]
      summary: Detalhe da despesa, com o payable vinculado quando futura (requer permission:view_expenses)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Expense" }
  /expenses/{id}/cancel:
    post:
      tags: [expenses]
      summary: >
        Cancela a despesa — só permitido antes de paga (RN 10.15.9); cancela também o
        Payable vinculado, se houver.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "201":
          description: Cancelada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Expense" }
        "409":
          description: Despesa já paga ou já cancelada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: false
      description: >
        TA-API-002 — chave opcional fornecida pelo cliente para evitar duplicar o efeito de um
        comando crítico em caso de reenvio (ex.: timeout de rede). Repetir a mesma chave com o
        mesmo payload devolve a resposta original; com payload diferente é tratado como uma nova
        operação, não como conflito.
      schema: { type: string, maxLength: 255 }
    SearchQuery:
      name: q
      in: query
      required: false
      description: Busca por nome (contains, case-insensitive).
      schema: { type: string }
    PageQuery:
      name: page
      in: query
      required: false
      description: >
        Presença deste parâmetro é o que ativa a paginação — sem ele a resposta é o array
        completo de sempre (retrocompatível com os seletores de produto/fornecedor/cliente
        usados em Compras/Vendas).
      schema: { type: integer, minimum: 1 }
    PageSizeQuery:
      name: pageSize
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  schemas:
...
    Product:
      type: object
      properties:
        id: { type: string, format: uuid }
        companyId: { type: string, format: uuid }
        sku: { type: string }
        name: { type: string }
        unit: { type: string }
        status: { type: string, enum: [active, inactive] }
        categoryId: { type: string, format: uuid, nullable: true }
        brandId: { type: string, format: uuid, nullable: true }
        variants:
          type: array
          description: Variantes operacionais do produto; usar id em compras, estoque e vendas.
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              skuVariant: { type: string }
        prices:
          type: array
          description: Na listagem contém somente o preço vigente; no detalhe contém o histórico completo, do mais recente ao mais antigo.
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              salePrice: { type: string }
              costPrice: { type: string, description: Calculado; nunca editável diretamente. }
              effectiveFrom: { type: string, format: date-time }
    PaginatedProducts:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Product" } }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
...
    Customer:
      type: object
      properties:
        id: { type: string, format: uuid }
        companyId: { type: string, format: uuid }
        name: { type: string }
        whatsapp: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
        instagram: { type: string, nullable: true }
        birthDate: { type: string, format: date, nullable: true }
        status: { type: string, enum: [active, inactive] }
    CustomerDetail:
      allOf:
        - $ref: "#/components/schemas/Customer"
        - type: object
          properties:
            summary:
              type: object
              properties:
                salesCount: { type: integer }
                totalPurchased: { type: string }
                averageTicket: { type: string }
                lastPurchaseAt: { type: string, format: date-time, nullable: true }
                productsPurchased: { type: integer }
                outstandingBalance: { type: string }
            recentSales:
              type: array
              maxItems: 5
              items:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  status: { type: string }
                  total: { type: string }
                  createdAt: { type: string, format: date-time }
    PaginatedCustomers:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Customer" } }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
...
    Expense:
      type: object
      properties:
        id: { type: string, format: uuid }
        companyId: { type: string, format: uuid }
        description: { type: string }
        category: { type: string }
        amount: { type: string }
        competenceDate: { type: string, format: date }
        dueDate: { type: string, format: date, nullable: true }
        paidAt: { type: string, format: date-time, nullable: true }
        financialAccountId: { type: string, format: uuid, nullable: true }
        payableId: { type: string, format: uuid, nullable: true }
        status: { type: string, enum: [pending, paid, cancelled] }
...
    ErrorResponse:
      type: object
      description: Formato padrão TA-API-001 (docs/architecture/overview.md, secao 10.2)
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            field: { type: string }
            details: { type: object }